Move SU, DG upwind and poisson bracket term generators into common module.
