+++ scripts/common/__init__.py
@@ -1,3 +1,11 @@
 from .meshes import set_up_mesh
 from .io import read_yaml_config, set_default_param
-from .rr import read_rr_config, rr_src_term, rr_steady_state
+from .rr import (
+    read_rr_config,
+    rr_DG_upwind_term,
+    rr_ExB_vel,
+    rr_src_term,
+    rr_steady_state,
+    rr_SU_term,
+)
+from .terms import poisson_bracket
